factored out data fetch and dnd handler
